Prevent Oblivion Knights casting Iron Maiden
This change was hardcoded in 1.13 but will now apply to previous versions of the game as well.

Credit to talonrage for the fix:
https://www.d2mods.info/forum/viewtopic.php?f=4&t=24136

The original Iron Maiden skill has been nerfed in Skills.txt, which is the one used by Oblivion Knights. The Necromancer now uses Iron Maiden2 instead, which is a copy of the original un-nerfed skill.

Any references to the Iron Maiden skill (or its skill ID, 76) must be updated to use the new Iron Maiden2 skill (skill ID 358).